refactor: Move icons, export independently
