+++ libs/novui/src/components/NovuiProvider.tsx
@@ -1,6 +1,6 @@
 import { MantineProvider } from '@mantine/core';
 import { FC, PropsWithChildren } from 'react';
-import { IconProvider } from './icons/IconProvider';
+import { IconProvider } from '../icons/IconProvider';
 type INovuiProviderProps = PropsWithChildren;
 
 /** Used to export a v7 Mantine provider */
