+++ libs/novui/src/icons/Icon.const.ts
