+++ libs/novui/src/icons/Icon.stories.tsx
@@ -4,8 +4,8 @@ import { DEFAULT_ICON_SIZE } from './Icon.const';
 import { IIconProps, IconSize } from './Icon.types';
 import { Grid, Group, ActionIcon, Code, Input } from '@mantine/core';
 import * as allIcons from './icon-registry';
-import { styled, VStack } from '../../../styled-system/jsx';
-import { text } from '../../../styled-system/recipes';
+import { styled, VStack } from '../../styled-system/jsx';
+import { text } from '../../styled-system/recipes';
 
 export default {
   title: 'Icons/Icons',
