+++ libs/novui/src/icons/IconProvider.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren } from 'react';
 import { IconContext } from 'react-icons';
 import { DEFAULT_ICON_SIZE } from './Icon.const';
-import { css } from '../../../styled-system/css';
+import { css } from '../../styled-system/css';
 
 const iconClassName = css({
   verticalAlign: 'middle',
