+++ libs/novui/src/icons/icon-registry.ts
