+++ libs/novui/src/icons/react-icons-md.d.ts
